Simplify reports page structure in PHP.
