+++ src/Menu/Reports.php
@@ -1,17 +1,17 @@
 <?php
 declare( strict_types=1 );
 
-namespace Automattic\WooCommerce\GoogleListingsAndAds\Menu\Reports;
+namespace Automattic\WooCommerce\GoogleListingsAndAds\Menu;
 
 use Automattic\WooCommerce\GoogleListingsAndAds\Infrastructure\Registerable;
 use Automattic\WooCommerce\GoogleListingsAndAds\Infrastructure\Service;
 
 /**
- * Class Products
+ * Class Reports
  *
  * @package Automattic\WooCommerce\GoogleListingsAndAds\Menu\Reports
  */
-class Products implements Service, Registerable {
+class Reports implements Service, Registerable {
 
 	/**
 	 * Register a service.
@@ -22,13 +22,13 @@ class Products implements Service, Registerable {
 			function() {
 				wc_admin_register_page(
 					[
-						'title'    => __( 'Products Report', 'google-listings-and-ads' ),
-						'parent'   => 'google-reports',
-						'path'     => '/google/reports/products',
-						'id'       => 'google-reports-products',
+						'title'    => __( 'Reports', 'google-listings-and-ads' ),
+						'parent'   => 'google-listings-and-ads-category',
+						'path'     => '/google/reports',
+						'id'       => 'google-reports',
 						'nav_args' => [
 							'order'  => 20,
-							'parent' => 'google-reports',
+							'parent' => 'google-listings-and-ads-category',
 						],
 					]
 				);
