--- src/Menu/Reports/Programs.php
@@ -1,38 +1,0 @@
-<?php
-declare( strict_types=1 );
-
-namespace Automattic\WooCommerce\GoogleListingsAndAds\Menu\Reports;
-
-use Automattic\WooCommerce\GoogleListingsAndAds\Infrastructure\Registerable;
-use Automattic\WooCommerce\GoogleListingsAndAds\Infrastructure\Service;
-
-/**
- * Class Programs
- *
- * @package Automattic\WooCommerce\GoogleListingsAndAds\Menu\Reports
- */
-class Programs implements Service, Registerable {
-
-	/**
-	 * Register a service.
-	 */
-	public function register(): void {
-		add_action(
-			'admin_menu',
-			function() {
-				wc_admin_register_page(
-					[
-						'title'    => __( 'Programs Report', 'google-listings-and-ads' ),
-						'parent'   => 'google-reports',
-						'path'     => '/google/reports/programs',
-						'id'       => 'google-reports-programs',
-						'nav_args' => [
-							'order'  => 10,
-							'parent' => 'google-reports',
-						],
-					]
-				);
-			}
-		);
-	}
-}
